sde_hdcp: typecast switch labels into int

+++ drivers/gpu/drm/msm/sde_hdcp_2x.c
@@ -423,11 +423,11 @@ static u8 sde_hdcp_2x_stream_type(u8 min_enc_level)
 	u8 const stream_type_0 = 0, stream_type_1 = 1;
 
 	switch (min_enc_level) {
-	case hdcp_min_enc_level_0:
-	case hdcp_min_enc_level_1:
+	case (int) hdcp_min_enc_level_0:
+	case (int) hdcp_min_enc_level_1:
 		stream_type = stream_type_0;
 		break;
-	case hdcp_min_enc_level_2:
+	case (int) hdcp_min_enc_level_2:
 		stream_type = stream_type_1;
 		break;
 	default:
